Fix error in deployment docs (#425)
* Fix error in deployment docs

Fixes an error in the deployment documentation and updates the text about certificates for combiners.

* Update deployment.rst

I removed the sections about certs and tokens. They interrupted the flow since they came before actual deployment of services. Instead we should have a link at the end, or under the combiner heading to a page / section that  provides an updated description of deploying in secure mode / using tokens.